backend/requirements.txt
# Core backend dependencies
psycopg2-binary==2.9.9
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0

# Data processing and ML
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
xgboost==2.0.3

# Time series analysis
statsmodels==0.14.1

# File processing
openpyxl==3.1.2
xlrd==2.0.1

# Redis for caching (optional)
redis==5.0.1

# Validation and serialization
zod==0.22.0  # For Python equivalent validation

# Development dependencies
pytest==7.4.4
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0